chore(gitbbon-search): change model dtype from fp32 to fp16
- Update all model loading configurations to use fp16 instead of fp32
- Reduces memory usage while maintaining embedding quality
- No quantization applied (vectors stored as Float32)

Files changed:
- src/modelHost.ts
- src/services/modelService.ts
- src/webview/e5.worker.ts
- src/webview/modelService.ts

+++ extensions/gitbbon-search/src/modelHost.ts
@@ -114,7 +114,7 @@ class ModelHost {
 			// "Expression produces a union type that is too complex to represent"
 			this.extractor = await (pipeline as Function)('feature-extraction', MODEL_NAME, {
 				device: useWebGPU ? 'webgpu' : 'wasm',
-				dtype: 'fp32',
+				dtype: 'fp16',
 				progress_callback: (p: { progress?: number; status?: string; file?: string }) => {
 					console.log('[ModelHost] Download progress:', p);
 					if (typeof p?.progress === 'number') {
